add docs and move types to types.rs
